still refactoring
only calender and tasklist needs refaktoring
