+++ lib/app_pages.dart
@@ -1,5 +1,5 @@
 export './features/focus/presentation/focus_page.dart';
-export './features/home/presentation/home_page.dart';
+export 'features/home/presentation/home_page.dart';
 export './features/settings/presentation/settings_page.dart';
 export './features/stats/presentation/stats_page.dart';
 export './features/tasks/presentation/tasks_page.dart';
