+++ lib/features/focus/application/application.dart
@@ -1,1 +1,2 @@
 export 'focus_summery_provider.dart';
+export 'stop_watcher.dart';
