+++ lib/features/focus/application/stop_watcher.dart
