+++ lib/features/focus/presentation/presentation.dart
@@ -1,3 +1,2 @@
 export 'focus_page.dart';
-export 'stop_watcher.dart';
 export 'focus_piechart.dart';
