+++ lib/features/home/presentation/home_page.dart
@@ -1,5 +1,6 @@
 import 'package:timety/commons.dart';
 import 'package:timety/features/focus/application/focus_summery_provider.dart';
+import 'package:timety/features/focus/presentation/focus_piechart.dart';
 
 
 class HomePage extends ConsumerWidget {
