+++ lib/features/tasks/application/application.dart
@@ -1,0 +1,1 @@
+export 'task_list_notifier.dart';
