+++ lib/features/tasks/domain/domain.dart
@@ -1,3 +1,2 @@
 export 'event.dart';
-export 'scheduler.dart';
 export 'task_item.dart';
