--- lib/features/tasks/domain/scheduler.dart
@@ -1,5 +1,0 @@
-import 'package:timety/commons.dart';
-
-class Scheduler {
-  // I have no idea how we are gonna do this
-}
