--- lib/widgets/_widgets.dart
@@ -1,6 +1,0 @@
-// Custom widgets, aka reusable components for the app
-
-export '../features/tasks/presentation/task_widget.dart';
-export '../features/focus/presentation/focus_piechart.dart';
-export '../features/tasks/presentation/events_list.dart';
-export '../features/tasks/presentation/calendar_widget.dart';
